[vusb] Domain ID not being set in device object.
There was a change to the vusb_daemon in the space between when it was forked
off the ctxusb_daemon and when it went into OpenXT. In this fix, the setting
of the device->VM field was lost. This problem caused further binding calls
for interfaces for DOM0 devices to not occur and thus prevented udev from
completing its probe operations. This is why all the other nodes (/dev/sxx,
 /sys/class/bsg, etc) were not created and why the problem is not seen in
XCXT 4.0.

OXT-73

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>